% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{acceptance_rate_plots}
\alias{acceptance_rate_plots}
\title{Acceptance rate comparison between results}
\usage{
acceptance_rate_plots(
  hier1,
  hier2,
  colours = c("blue", "green"),
  time = NA,
  hierarchical
)
}
\arguments{
\item{hier1}{hierarchical/progressive Monte Carlo fusion result 
(will be in blue on plots by default)}

\item{hier2}{hierarchical/progressive Monte Carlo fusion result 
(will be in green on plots by default)}

\item{colours}{vector of 2 colours, where colours[1] is for hier1 and 
colours[2] is for hier2}

\item{time}{time that you are comparing, if not appropriate, can use NA}

\item{hierarchical}{logical value to determine if it was a result from 
                    hierarchical fusion (TRUE) or progressive fusion (FALSE)

return None}
}
\description{
Creates plots to compare the acceptance probabilities and time taken to run
}
