% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_fusion.R
\name{bal_binary_fusion_SMC_uniGaussian}
\alias{bal_binary_fusion_SMC_uniGaussian}
\title{(Balanced Binary) D&C Monte Carlo Fusion using SMC}
\usage{
bal_binary_fusion_SMC_uniGaussian(
  N_schedule,
  m_schedule,
  time_schedule,
  base_samples,
  L,
  mean,
  sd,
  start_beta,
  precondition = TRUE,
  resampling_method = "multi",
  ESS_threshold = 0.5,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{N_schedule}{vector of length (L-1), where N_schedule[l] is the number 
of samples per node at level l}

\item{m_schedule}{vector of length (L-1), where m_schedule[l] is the number 
of samples to fuse for level l}

\item{time_schedule}{vector of length(L-1), where time_schedule[l] is the time 
chosen for Fusion at level l}

\item{base_samples}{list of length (1/start_beta), where base_samples[[c]] 
contains the samples for the c-th node in the level}

\item{L}{total number of levels in the hierarchy}

\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{start_beta}{beta for the base level}

\item{precondition}{either a logical value to determine if preconditioning values are 
used (TRUE - and is set to be the variance of the sub-posterior samples)
or not (FALSE - and is set to be 1 for all sub-posteriors),
or a list of length (1/start_beta) where precondition[[c]]
is the preconditioning value for sub-posterior c. Default is TRUE}

\item{resampling_method}{method to be used in resampling, default is multinomial 
resampling ('multi'). Other choices are stratified 
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{ESS_threshold}{number between 0 and 1 defining the proportion 
of the number of samples that ESS needs to be
lower than for resampling (i.e. resampling is carried 
out only when ESS < N*ESS_threshold)}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
A list with components:
\describe{
  \item{particles}{list of length (L-1), where particles[[l]][[i]] are the
                   particles for level l, node i}
  \item{proposed_samples}{list of length (L-1), where proposed_samples[[l]][[i]]
                          are the proposed samples for level l, node i}
  \item{time}{list of length (L-1), where time[[l]][[i]] is the run time for level l,
              node i}
  \item{ESS}{list of length (L-1), where ESS[[l]][[i]] is the effective 
             sample size of the particles after each step BEFORE deciding 
             whether or not to resample for level l, node i}
  \item{CESS}{list of length (L-1), where CESS[[l]][[i]] is the conditional
              effective sample size of the particles after each step}
  \item{resampled}{list of length (L-1), where resampled[[l]][[i]] is a 
                   boolean value to record if the particles were resampled
                   after each step; rho and Q for level l, node i}
  \item{precondition_values}{preconditioning values used in the algorithm 
                            for each node}
  \item{diffusion_times}{vector of length (L-1), where diffusion_times[l]
                        are the times for fusion in level l}
}
}
\description{
(Balanced Binary) D&C Monte Carlo Fusion with univariate Gaussian target
}
