% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_sampler.R
\name{base_rejection_sampler_exp_4}
\alias{base_rejection_sampler_exp_4}
\title{Rejection sampler for base level}
\usage{
base_rejection_sampler_exp_4(
  beta,
  mean = 0,
  nsamples,
  proposal_mean,
  proposal_sd,
  dominating_M
)
}
\arguments{
\item{beta}{real value between 0 and 1}

\item{mean}{mean value, defaults to 0}

\item{nsamples}{number of samples per node}

\item{proposal_mean}{mean of proposal density for rejection sampler}

\item{proposal_sd}{st.dev of proposal density for rejection sampler}

\item{dominiating_M}{constant M to bound the target density}
}
\value{
return_name return description
}
\description{
Sample for base level (samples of exp(-(beta*(x-mean)^4)/2))
}
\examples{
base_rejection_sampler_exp_4(beta = 1/2, 
                             nsamples = 100000, 
                             proposal_mean = 0, 
                             proposal_sd = 1, 
                             dominating_M = 1.3)
                             
}
