% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_proposal_cov}
\alias{calculate_proposal_cov}
\title{Calculate the proposal covariance matrix}
\usage{
calculate_proposal_cov(time, weights)
}
\arguments{
\item{time}{T for fusion algorithm}

\item{weights}{list of m matrices of the same dimension}
}
\value{
proposal covariance matrix
}
\description{
Calculation of the proposal covariance matrix for Monte Carlo Fusion
}
\examples{
m1 <- matrix(c(1,2,3,4), nrow = 2, ncol = 2)
m2 <- matrix(c(5,6,7,8), nrow = 2, ncol = 2)
m3 <- matrix(c(9,10,11,12), nrow = 2, ncol = 2)
calculate_proposal_cov(time = 0.5, weights = list(m1, m2, m3))
}
