% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixG_sampler.R
\name{dnorm_mix}
\alias{dnorm_mix}
\title{Density of a mixture Gaussian}
\usage{
dnorm_mix(x, n_comp, weights, means, sds)
}
\arguments{
\item{x}{real value}

\item{n_comp}{integer number of components of mixture Gaussian}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gassuan}

\item{sds}{vector: st.devs of mixture Gaussian}
}
\value{
value: value of a mixture Gaussian at given point x
}
\description{
Returns the value of a mixture Gaussian at given point x
}
\examples{
weights <- c(0.4, 0.6)
means <- c(-3, 6)
sds <- c(1 ,2)
curve(dnorm_mix(x,
                n_comp = n_comp,
                weights = weights,
                means = means,
                sds = sds),
      -10, 15,
      ylim = c(0, 0.2), ylab = 'pdf')

}
