% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixG_sampler.R
\name{dnorm_mix_tempered}
\alias{dnorm_mix_tempered}
\title{Density of tempered mixture Gaussian}
\usage{
dnorm_mix_tempered(x, n_comp, weights, means, sds, beta, normalised = TRUE)
}
\arguments{
\item{x}{real value}

\item{n_comp}{integer number of components of mixture Gaussian}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gassuan}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{beta}{real value}

\item{normalised}{boolean value to determine if normalisation constant is calculated}
}
\value{
value: value of a normalised tempered mixture Gaussian at given point x
}
\description{
Returns the value of a normalised tempered mixture Gaussian at given point x
}
\examples{
weights <- c(0.4, 0.6)
means <- c(-3, 6)
sds <- c(1, 2)
beta <- 1/4
curve(dnorm_mix_tempered(x,
                         n_comp = 2,
                         weights = weights,
                         means = means,
                         sds = sds,
                         beta = beta,
                         normalised = FALSE),
      -15, 20, ylim = c(0, 0.8), ylab = 'pdf')
curve(dnorm_mix_tempered(x,
                         n_comp = 2,
                         weights = weights,
                         means = means,
                         sds = sds,
                         beta = beta,
                         normalised = TRUE),
      add = T, lty = 2)

}
