% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_sampler.R
\name{dnorm_tempered}
\alias{dnorm_tempered}
\title{Tempered Gaussian Distribution}
\usage{
dnorm_tempered(x, mean = 0, sd = 1, beta)
}
\arguments{
\item{x}{real value}

\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{beta}{real value}
}
\value{
density of tempered Gaussian distribution
}
\description{
Density of tempered Gaussian distribution
}
\examples{
# standard normal
curve(dnorm(x), -5, 5)
# standard normal to the power 1/2
curve(dnorm_tempered(x, mean = 0, sd = 1, beta = 1/2), -5, 5, add = T, col = 'blue')

}
