% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_fusion.R
\name{ea_biGaussian_DL_PT}
\alias{ea_biGaussian_DL_PT}
\title{Diffusion probability for the Exact Algorithm for langevin diffusion for
bivariate tempered Gaussian distribution}
\usage{
ea_biGaussian_DL_PT(
  x0,
  y,
  s,
  t,
  mean_vec,
  sd_vec,
  corr,
  beta,
  precondition_mat,
  transform_mats,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  logarithm
)
}
\arguments{
\item{x0}{start value (vector of length 2)}

\item{y}{end value (vector of length 2)}

\item{s}{start time}

\item{t}{end time}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{beta}{real value}

\item{precondition_mat}{precondition matrix}

\item{transform_mats}{list of transformation matrices where 
transform_mats$to_Z is the transformation matrix to Z space
and transform_mats$to_X is the transformation matrix to 
X space}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{logarithm}{logical value to determine if log probability is
returned (TRUE) or not (FALSE)}
}
\value{
acceptance probability of simulating langevin diffusion with pi =
        bivariate tempered Gaussian distribution
}
\description{
Simulate langevin diffusion using the Exact Algorithm where pi =
bivariate tempered Gaussian distribution
}
