% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRR_fusion.R
\name{ea_BRR_DL_PT}
\alias{ea_BRR_DL_PT}
\title{Diffusion probability for the Exact Algorithm for Langevin diffusion for
Bayesian logistic regression}
\usage{
ea_BRR_DL_PT(
  dim,
  x0,
  y,
  s,
  t,
  data,
  transformed_design_mat,
  nu,
  sigma,
  prior_means,
  prior_variances,
  C,
  precondition_mat,
  transform_mats,
  cv_location = "hypercube_centre",
  diffusion_estimator,
  beta_NB = 10,
  gamma_NB_n_points = 2,
  local_bounds = TRUE,
  logarithm
)
}
\arguments{
\item{dim}{dimension of the predictors (= p+1)}

\item{x0}{start value (vector of length dim)}

\item{y}{end value (vector of length dim)}

\item{s}{start time}

\item{t}{end time}

\item{data}{list of length 4 where data[[c]]$y is the vector for y responses 
and data[[c]]$X is the design matrix for the covariates for
sub-posterior c}

\item{nu}{degrees of freedom in t-distribution}

\item{sigma}{scale parameter in t-distribution}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{C}{overall number of sub-posteriors}

\item{precondition_mat}{precondition matrix}

\item{transform_mats}{list of transformation matrices where 
transform_mats$to_Z is the transformation matrix to Z space
and transform_mats$to_X is the transformation matrix to 
X space}

\item{cv_location}{string to determine what the location of the control variate
should be. Must be either 'mode' where the MLE estimator 
will be used or 'hypercube_centre' (default) to use the centre
of the simulated hypercube}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{local_bounds}{logical value indicating if local bounds for the phi function
are used (default is TRUE)}

\item{logarithm}{logical value to determine if log probability is
returned (TRUE) or not (FALSE)}
}
\value{
acceptance probability of simulating Langevin diffusion where target
        is the posterior for a logistic regression model with Gaussian priors
}
\description{
Simulate Langevin diffusion using the Exact Algorithm where target is the
posterior for a logistic regression model with Gaussian priors
}
