% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{ea_exp_4_DL}
\alias{ea_exp_4_DL}
\title{Exact Algorithm for langevin diffusion with pi = exp(-(beta*(x-mean)^4)/2)}
\usage{
ea_exp_4_DL(N, input_samples, time, mean, beta, precondition)
}
\arguments{
\item{N}{number of samples}

\item{input_samples}{input samples for the algorithm distributed according to
pi = exp(-(beta*(x-mean)^4)/2)}

\item{time}{time T for Exact Algorithm}

\item{mean}{mean value}

\item{beta}{real value}

\item{precondition}{precondition value (i.e the covariance for 
the Langevin diffusion)}
}
\value{
end points of the Exact Algorithm which should also be distributed
        according to pi = exp(-(beta*(x-mean)^4)/2)
}
\description{
Exact Algorithm with pi = exp(-(beta*(x-mean)^4)/2)
}
