% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixG_fusion.R
\name{ea_mixG_DL_PT}
\alias{ea_mixG_DL_PT}
\title{Exact Algorithm for Langevin diffusion with pi as a tempered mixture Gaussian}
\usage{
ea_mixG_DL_PT(
  x0,
  y,
  s,
  t,
  n_comp,
  weights,
  means,
  sds,
  beta,
  precondition,
  bounds_multiplier = 1.1,
  logarithm
)
}
\arguments{
\item{x0}{start value}

\item{y}{end value}

\item{s}{start time}

\item{t}{end time}

\item{n_comp}{integer number of components of mixture Gaussian}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gaussian}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{beta}{real value}

\item{precondition}{precondition value (i.e the covariance for 
the Langevin diffusion)}

\item{bounds_multiplier}{scalar value to multiply bounds by 
(should greater than or equal to 1)}

\item{logarithm}{logical value to determine if log probability is 
returned (TRUE) or not (FALSE)}
}
\description{
Simulate Langevin diffusion using the Exact Algorithm with pi = exp(-(beta*x^4)/2)
}
\examples{
weights <- c(0.4, 0.6)
means <- c(-8, 15)
sds <- c(1, 2)
beta <- 1/4
precondition <- 1.5
x0 <- -9
y <- -8
s <- 0
t <- 1
# simulate event probability of diffusion from -9 to -8 between [0,1]
ea_mixG_DL_PT(x0 = x0,
              y = y,
              s = s,
              t = t,
              n_comp = 2,
              weights = weights,
              means = means,
              sds = sds,
              beta = beta,
              precondition = precondition,
              bounds_multiplier = 1,
              logarithm = FALSE)
# simulate event probability of diffusion from -9 to 15 between [0,1]
x0 <- -9
y <- 15
ea_mixG_DL_PT(x0 = x0,
              y = y,
              s = s,
              t = t,
              n_comp = 2,
              weights = weights,
              means = means,
              sds = sds,
              beta = beta,
              precondition = precondition,
              bounds_multiplier = 1,
              logarithm = FALSE)

}
