% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_fusion.R
\name{ea_phi_biGaussian_DL}
\alias{ea_phi_biGaussian_DL}
\title{phi-function for bivariate tempered Gaussian distribution}
\usage{
ea_phi_biGaussian_DL(
  x,
  mean_vec,
  sd_vec,
  corr,
  beta,
  precondition_mat,
  transform_mat
)
}
\arguments{
\item{x}{vector of length 2}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{beta}{real value}

\item{precondition_mat}{precondition matrix}

\item{transform_mat}{the transformation matrix}
}
\value{
real value
}
\description{
phi-function for the Exact Algorithm for bivariate tempered Gaussian distribution
}
