% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_biGaussian_DL_LB}
\alias{ea_phi_biGaussian_DL_LB}
\title{Obtain the global lower bound for phi function}
\usage{
ea_phi_biGaussian_DL_LB(mean_vec, sd_vec, corr, beta, precondition_mat)
}
\arguments{
\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{beta}{real value}

\item{precondition_mat}{precondition matrix}

\item{x}{vector of length 2}

\item{transform_mat}{the transformation matrix}
}
\value{
The global lower bound of phi
}
\description{
Finds the global bound of the phi function between a given interval
}
