% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_exp_4_DL}
\alias{ea_phi_exp_4_DL}
\title{phi-function for exp(-(((x-mean)^4)*beta)/2)}
\usage{
ea_phi_exp_4_DL(x, mean, beta, precondition)
}
\arguments{
\item{x}{real value}

\item{mean}{mean value}

\item{beta}{beta value}

\item{precondition}{precondition value}
}
\value{
real value
}
\description{
phi-function for the Exact Algorithm for exp(-(((x-mean)^4)*beta)/2)
}
\examples{
curve(ea_phi_exp_4_DL(x, 0, 1, 1), from = -4, to = 4)
}
