% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_exp_4_DL_bounds}
\alias{ea_phi_exp_4_DL_bounds}
\title{Obtain bounds for phi function}
\usage{
ea_phi_exp_4_DL_bounds(mean, beta, precondition, lower, upper)
}
\arguments{
\item{mean}{mean value}

\item{beta}{beta value}

\item{precondition}{precondition value}

\item{lower}{lower end of interval}

\item{upper}{upper end of interval}
}
\value{
A list of components
\describe{
  \item{LB}{lower bound of phi}
  \item{UB}{upper bound of phi}
}
}
\description{
Finds the lower and upper bounds of the phi function between a given interval
}
\examples{
mu <- 0.435
beta <- 0.583
precondition <- 1.243
lower <- 0
upper <- 1.2

curve(ea_phi_exp_4_DL(x, mu, beta, precondition), lower, upper, ylab = 'phi')
abline(v=c(lower, upper))
abline(h=ea_phi_exp_4_DL_bounds(mean = mu,
                                beta = beta,
                                precondition = precondition,
                                lower = lower,
                                upper = upper)
       col = 'red', lty = 2)
}
