% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_mixG_DL}
\alias{ea_phi_mixG_DL}
\title{phi-function for a tempered mixture Gaussian}
\usage{
ea_phi_mixG_DL(x, n_comp, weights, means, sds, beta, precondition)
}
\arguments{
\item{x}{real value}

\item{n_comp}{integer number of components of mixture Gaussian}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gassuan}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{beta}{real value}

\item{precondition}{precondition value}
}
\value{
real value
}
\description{
phi-function for the Exact Algorithm for a tempered mixture Gaussian
}
\examples{
weights <- c(0.4, 0.6)
means <- c(-3, 6)
sds <- c(1, 2)
beta <- 1/4
precondition <- 1
curve(dnorm_mix_tempered(x,
                         n_comp = 2,
                         weights = weights,
                         means = means,
                         sds = sds,
                         beta = beta),
      -20, 40,  n = 100000, ylim = c(-0.5, 2), ylab = 'y')
curve(ea_phi_mixG_DL(x,
                     n_comp = 2,
                     weights = weights,
                     means = means,
                     sds = sds,
                     beta = beta,
                     precondition = precondition),
      add = T, n = 100000, col = 'red')
}
