% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_mixG_DL_LB}
\alias{ea_phi_mixG_DL_LB}
\title{Obtain the global lower bound for phi function}
\usage{
ea_phi_mixG_DL_LB(
  n_comp,
  weights,
  means,
  sds,
  beta,
  precondition,
  bounds_multiplier
)
}
\arguments{
\item{n_comp}{integer number of components of mixture Gaussian}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gassuan}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{beta}{real value}

\item{normalised}{boolean value to determine if normalisation constant is calculated}
}
\value{
The global lower bound of phi
}
\description{
Finds the global bound of the phi function between a given interval
}
\examples{
weights <- c(0.4, 0.6)
means <- c(-3, 6)
sds <- c(1, 2)
beta <- 1
precondition <- 1
curve(ea_phi_mixG_DL(x,
                     n_comp = 2,
                     weights = weights,
                     means = means,
                     sds = sds,
                     beta = beta,
                     precondition = precondition),
      from = -20, to = 25, n = 10000,
      ylab = 'phi')
PHI <- ea_phi_mixG_DL_LB(n_comp = 2,
                         weights = weights,
                         means = means,
                         sds = sds,
                         beta = beta,
                         precondition = precondition,
                         bounds_multiplier = 1)
abline(h = PHI, col = 'red')
}
