% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_Gaussian_fusion.R
\name{ea_phi_multiGaussian_DL}
\alias{ea_phi_multiGaussian_DL}
\title{phi-function for multivariate tempered Gaussian distribution}
\usage{
ea_phi_multiGaussian_DL(dim, x, mu, inv_Sigma, beta, precondition_mat)
}
\arguments{
\item{dim}{dimension}

\item{x}{vector of length dim}

\item{mu}{vector of length dim for mean}

\item{inv_Sigma}{dim x dim inverse covariance matrix}

\item{beta}{real value}

\item{precondition_mat}{precondition matrix}
}
\value{
real value
}
\description{
phi-function for the Exact Algorithm for multivariate tempered Gaussian distribution
}
