% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_multiGaussian_DL_bounds}
\alias{ea_phi_multiGaussian_DL_bounds}
\title{Obtain bounds for phi function}
\usage{
ea_phi_multiGaussian_DL_bounds(
  mu,
  inv_Sigma,
  beta,
  precondition_mat,
  hypercube_vertices
)
}
\arguments{
\item{mu}{vector of length dim for mean}

\item{inv_Sigma}{dim x dim inverse covariance matrix}

\item{beta}{real value}

\item{precondition_mat}{dim x dim precondition matrix}

\item{hypercube_vertices}{list with item named "V" to determine the points
to evaluate phi which give the bounds of phi}
}
\value{
A list of components
\describe{
  \item{LB}{lower bound of phi}
  \item{UB}{upper bound of phi}
}
}
\description{
Finds the lower and upper bounds of the phi function between an interval
}
