% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_multiGaussian_DL_LB}
\alias{ea_phi_multiGaussian_DL_LB}
\title{Obtain the global lower bound for phi function}
\usage{
ea_phi_multiGaussian_DL_LB(mu, inv_Sigma, beta, precondition_mat)
}
\arguments{
\item{mu}{vector of length dim for mean}

\item{inv_Sigma}{dim x dim inverse covariance matrix}

\item{beta}{real value}

\item{precondition_mat}{dim x dim precondition matrix}
}
\value{
The global lower bound of phi
}
\description{
Finds the global bound of the phi function between a given interval
}
