% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_uniGaussian_DL}
\alias{ea_phi_uniGaussian_DL}
\title{phi-function for tempered Gaussian distribution}
\usage{
ea_phi_uniGaussian_DL(x, mean, sd, beta, precondition)
}
\arguments{
\item{x}{real value}

\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{beta}{real value}

\item{precondition}{precondition value}
}
\value{
real value
}
\description{
phi-function for the Exact Algorithm for tempered Gaussian distribution
}
\examples{
curve(ea_phi_uniGaussian_DL(x, 0, 1, 1, 1), from = -4, to = 4)
}
