% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_uniGaussian_DL_bounds}
\alias{ea_phi_uniGaussian_DL_bounds}
\title{Obtain bounds for phi function}
\usage{
ea_phi_uniGaussian_DL_bounds(mean, sd, beta, precondition, lower, upper)
}
\arguments{
\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{beta}{real value}

\item{precondition}{precondition value}

\item{lower}{lower end of interval}

\item{upper}{upper end of interval}
}
\value{
A list of components
\describe{
  \item{LB}{lower bound of phi}
  \item{UB}{upper bound of phi}
}
}
\description{
Finds the lower and upper bounds of the phi function between an interval
}
\examples{
mu <- 0.423
sd <- 3.231
beta <- 0.8693
precondition <- 1.243
lower <- -2.823
upper <- 4.322
curve(ea_phi_uniGaussian_DL(x, mu, sd, beta, precondition), lower, upper)
abline(h=ea_phi_uniGaussian_DL_bounds(mean = mu,
                                      sd = sd,
                                      beta = beta,
                                      lower = lower,
                                      upper = upper,
                                      precondition = precondition),
       col = 'red', lty = 2)

# another example where the mean is not in the interval
mu <- 0.423
sd <- 3.231
beta <- 0.8693
precondition <- 1.243
lower <- 2.823
upper <- 5.322
curve(ea_phi_uniGaussian_DL(x, mu, sd, beta, precondition), lower, upper)
abline(h=ea_phi_uniGaussian_DL_bounds(mean = mu,
                                      sd = sd,
                                      beta = beta,
                                      precondition = precondition,
                                      lower = lower,
                                      upper = upper),
       col = 'red', lty = 2)
}
