% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ea_phi_uniGaussian_DL_LB}
\alias{ea_phi_uniGaussian_DL_LB}
\title{Obtain the global lower bound for phi function}
\usage{
ea_phi_uniGaussian_DL_LB(mean, sd, beta, precondition)
}
\arguments{
\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{beta}{real value}

\item{precondition}{precondition value}
}
\value{
The global lower bound of phi
}
\description{
Finds the global bound of the phi function between a given interval
}
\examples{
mu <- 0.423
sd <- 3.231
beta <- 0.8693
precondition <- 1.243
lower <- -2.823
upper <- 4.322
curve(ea_phi_uniGaussian_DL(x, mu, sd, beta, precondition), lower, upper)
abline(h=ea_phi_uniGaussian_DL_bounds(mean = mu,
                                      sd = sd,
                                      beta = beta,
                                      precondition = precondition,
                                      lower = lower,
                                      upper = upper),
       col = 'red', lty = 2)
abline(h=ea_phi_uniGaussian_DL_LB(mean = mu,
                                  sd = sd,
                                  beta = beta, 
                                  precondition = precondition),
       col = 'blue', lty = 3)

# another example where the mean is not in the interval
mu <- 0.423
sd <- 3.231
beta <- 0.8693
precondition <- 1.243
lower <- 2.823
upper <- 5.322
curve(ea_phi_uniGaussian_DL(x, mu, sd, beta, precondition), lower, upper)
abline(h=ea_phi_uniGaussian_DL_bounds(mean = mu,
                                      sd = sd,
                                      beta = beta,
                                      lower = lower,
                                      upper = upper,
                                      precondition = precondition),
       col = 'red', lty = 2)
abline(h=ea_phi_uniGaussian_DL_LB(mean = mu,
                                  sd = sd,
                                  beta = beta, 
                                  precondition = precondition),
       col = 'blue', lty = 3)
}
