% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_fusion.R
\name{ea_uniGaussian_DL}
\alias{ea_uniGaussian_DL}
\title{Exact Algorithm for Langevin diffusion for tempered Gaussian distribution}
\usage{
ea_uniGaussian_DL(N, input_samples, time, mean, sd, beta, precondition)
}
\arguments{
\item{N}{number of samples}

\item{input_samples}{input samples for the algorithm distributed according to
pi = exp(-(beta*(x-mean)^4)/2)}

\item{time}{time T for Exact Algorithm}

\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{beta}{real value}

\item{precondition}{precondition value (i.e the covariance for 
the Langevin diffusion)}
}
\value{
end points of the Exact Algorithm which should also be distributed
        according to pi = tempered Gaussian distribution
}
\description{
Exact Algorithm with pi = tempered Gaussian distribution
}
