% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Euclidean_distance}
\alias{Euclidean_distance}
\title{Euclidean distance between two vectors}
\usage{
Euclidean_distance(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
the Euclidean distance between vectors x and y
}
\description{
Calculates the Euclidean distance between two vectoirs
}
\examples{
x <- c(0.3, 0.2, 0.5, 1.2)
y <- c(-0.5, 0.8, 1.4, 0.9)
Euclidean_distance(x, y)
# should equal:
sqrt(0.8^2 + 0.6^2 + 0.9^2 + 0.3^2)
}
