% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_sampler.R
\name{exp_4_density}
\alias{exp_4_density}
\title{Tempered target density exp(-(((x-mean)^4)*beta)/2)}
\usage{
exp_4_density(x, mean = 0, beta = 1)
}
\arguments{
\item{x}{real value}

\item{mean}{mean value, defaults to 0}

\item{beta}{real value between 0 and 1, defaults to 1}
}
\value{
real value
}
\description{
Normalised tempered target density
}
\examples{
curve(exp_4_density(x, 1/2), from = -4, to = 4)

}
