% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_fusion.R
\name{fusion_biGaussian}
\alias{fusion_biGaussian}
\title{Preconditioned Monte Carlo Fusion [on a single core]}
\usage{
fusion_biGaussian(
  N,
  m,
  time,
  samples_to_fuse,
  mean_vec,
  sd_vec,
  corr,
  betas,
  precondition_matrices,
  inv_precondition_matrices,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2
)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containing
the samples for the c-th sub-posterior}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{betas}{vector of length m, where betas[c] is the inverse temperature (beta)
for c-th sub-posterior (can also pass in one number if they are all
at the same inverse temperature)}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{precondition_matricies}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}
}
\value{
samples: fusion samples

iters_rho: number of iterations from the first accept/reject step (rho)

iters_Q: number of iterations from the second (diffusion) accept/reject step (Q)
}
\description{
Monte Carlo Fusion for sub-posteriors that are tempered bivariate Gaussian distributions
}
