% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{fusion_exp_4}
\alias{fusion_exp_4}
\title{Preconditioned Monte Carlo Fusion [on a single core]}
\usage{
fusion_exp_4(
  N,
  m,
  time,
  samples_to_fuse,
  mean,
  betas,
  precondition_values,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2
)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containing 
the samples for the c-th sub-posterior}

\item{mean}{mean value of target distribution}

\item{betas}{vector of length m, where betas[c] is the inverse temperature 
(beta) for c-th sub-posterior (can also pass in one number if 
they are all at the same inverse temperature)}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}
}
\value{
A list with components:
\describe{
  \item{samples}{fusion samples}
  \item{iters_rho}{number of iterations for rho step}
  \item{iters_Q}{number of iterations for Q step}
}
}
\description{
Monte Carlo Fusion for sub-posteriors of the form exp(-((x^4)*beta)/2)
}
