% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_Gaussian_fusion.R
\name{fusion_multiGaussian}
\alias{fusion_multiGaussian}
\title{Generalised Monte Carlo Fusion (rejection sampling) [on a single core]}
\usage{
fusion_multiGaussian(
  N,
  m,
  time,
  samples_to_fuse,
  dim,
  mu,
  inv_Sigma,
  betas,
  precondition_matrices,
  inv_precondition_matrices
)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containing
the samples for the c-th sub-posterior}

\item{dim}{dimension}

\item{mu}{vector of length dim for mean}

\item{inv_Sigma}{dim x dim inverse covariance matrix}

\item{betas}{vector of length m, where betas[c] is the inverse temperature (beta)
for c-th sub-posterior (can also pass in one number if they are all
at the same inverse temperature)}

\item{precondition_matrices}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}
}
\value{
samples: fusion samples

iters_rho: number of iterations from the first accept/reject step (rho)

iters_Q: number of iterations from the second (diffusion) accept/reject step (Q)
}
\description{
Generalised Monte Carlo Fusion with multivariate Gaussian target
}
