% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_rho_univariate}
\alias{log_rho_univariate}
\title{Calculate the logarithm of rho (univariate)}
\usage{
log_rho_univariate(x, weighted_mean, time, precondition_values)
}
\arguments{
\item{x}{vector of sampled sub-posterior values}

\item{weighted_mean}{weighted mean of sampled sub-posterior values}

\item{time}{time T for fusion algorithm}

\item{precondition_values}{precondition values associated to each sub-posterior}
}
\value{
the logarithm of rho
}
\description{
Calculation of the log of rho acceptance probability or weight when target is
univariate
}
\examples{
x <- rnorm(4, 0, 1)
precondition_vals <- c(1, 2, 3, 4)
weighted_mean <- weighted_mean_univariate(x = x,
                                          weights = 1/precondition_vals)
log_rho_univariate(x = x,
                   weighted_mean = weighted_mean,
                   time = 0.5,
                   precondition_values = precondition_vals)
}
