% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logsumexp}
\alias{logsumexp}
\title{Calculate the logarithm of the sum of the exponentials of the arguments}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{vector}
}
\value{
the logarithm of the sum of the exponentials of elements of x
        i.e. returns log(sum(exp(x)))
}
\description{
Calculation of the log of the sum of exponential of x, but avoids computational 
underflow / overflow
}
\examples{
x <- c(1000.01, 1000.02)
y1 <- log(sum(exp(x)))
print(y1)
y2 <- logsumexp(x)
print(y2) 
}
