% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maximal_distance_hypercube_to_cv}
\alias{maximal_distance_hypercube_to_cv}
\title{Maximal distance to hypercube}
\usage{
maximal_distance_hypercube_to_cv(
  beta_hat,
  hypercube_vertices,
  transform_to_X,
  transform_to_Z,
  hypercube_centre = FALSE
)
}
\arguments{
\item{beta_hat}{vector}

\item{transform_to_X}{transformation matrix to Z-space (transformed space)}

\item{hypercube}{vertices matrix of hypercube vertices}
}
\value{
The maximal distance from beta_hat to point in hypercube
}
\description{
Calculates the maximal distance from a vector to a hypercube
}
