% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Simulate from a Multivariate Gaussian Distribution}
\usage{
mvrnormArma(N, mu, Sigma)
}
\arguments{
\item{N}{the number of samples required}

\item{mu}{a vector of mean values}

\item{Sigma}{positive-definite symmetric matrix specifying the covariance of variables}
}
\value{
samples from a multivariate Gaussian distribution
}
\description{
Produces samples from the specified multivariate Gaussian distribution
}
\examples{
samples <- mvrnormArma(N = 10000, mu = c(0, 0), Sigma = diag(2))
}
