% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma_tempered}
\alias{mvrnormArma_tempered}
\title{Simulate from a tempered Multivariate Gaussian Distribution}
\usage{
mvrnormArma_tempered(N, mu, Sigma, beta)
}
\arguments{
\item{N}{the number of samples required}

\item{mu}{a vector of mean values}

\item{Sigma}{positive-definite symmetric matrix specifying the covariance of variables}

\item{beta}{inverse temperature}
}
\value{
samples from a tempered multivariate Gaussian distribution
}
\description{
Produces samples from the specified tempered multivariate Gaussian distribution
}
\examples{
samples <- mvrnormArma(N = 10000,
                       mu = c(0, 0),
                       Sigma = diag(2))
tempered_samples <- mvrnormArma_tempered(N = 10000,
                                         mu = c(0, 0),
                                         Sigma = diag(2),
                                         beta = 2)
}
