% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{parallel_fusion_exp_4}
\alias{parallel_fusion_exp_4}
\title{Preconditioned Monte Carlo Fusion [parallel]}
\usage{
parallel_fusion_exp_4(
  N,
  m,
  time,
  samples_to_fuse,
  mean,
  betas,
  precondition_values,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{N}{number of sample}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containing 
the samples for the c-th sub-posterior}

\item{mean}{mean value of target distribution}

\item{betas}{vector of length m, where betas[c] is the inverse temperature 
(beta) for c-th sub-posterior (can also pass in one number if 
they are all at the same inverse temperature)}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
A list with components:
\describe{
  \item{samples}{fusion samples}
  \item{rho}{acceptance rate for rho step}
  \item{Q}{acceptance rate for Q step}
  \item{rhoQ}{overall acceptance rate}
  \item{time}{run-time of fusion sampler}
  \item{rho_iterations}{number of iterations for rho step}
  \item{Q_iterations}{number of iterations for Q step}
  \item{precondition_values}{list of length 2 where precondition_values[[2]] 
                             are the pre-conditioning values that were used 
                             and precondition_values[[1]] are the combined 
                             precondition values}
}
}
\description{
Monte Carlo Fusion for sub-posteriors of the form exp(-beta*((x-mean)^4)/2)
}
