% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_fusion.R
\name{parallel_fusion_SMC_biGaussian}
\alias{parallel_fusion_SMC_biGaussian}
\title{Generalised Monte Carlo Fusion [parallel]}
\usage{
parallel_fusion_SMC_biGaussian(
  particles_to_fuse,
  N,
  m,
  time,
  mean_vec,
  sd_vec,
  corr,
  betas,
  precondition_matrices,
  resampling_method = "multi",
  ESS_threshold = 0.5,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{particles_to_fuse}{list of length m, where particles_to_fuse[[c]]
contains the particles for the c-th sub-posterior
(a list of particles to fuse can be initialised by 
initialise_particle_sets() function)}

\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{precondition_matrices}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}

\item{resampling_method}{method to be used in resampling, default is multinomial 
resampling ('multi'). Other choices are stratified 
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{ESS_threshold}{number between 0 and 1 defining the proportion 
of the number of samples that ESS needs to be
lower than for resampling (i.e. resampling is carried 
out only when ESS < N*ESS_threshold)}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
A list with components:
\describe{
  \item{particles}{particles returned from fusion sampler}
  \item{proposed_samples}{proposal samples from fusion sampler}
  \item{time}{run-time of fusion sampler}
  \item{ESS}{list of length (L-1), where ESS[[l]][[i]] is the effective 
             sample size of the particles after each step BEFORE deciding 
             whether or not to resample for level l, node i}
  \item{CESS}{list of length (L-1), where CESS[[l]][[i]] is the conditional
              effective sample size of the particles after each step}
  \item{resampled}{list of length (L-1), where resampled[[l]][[i]] is a 
                   boolean value to record if the particles were resampled
                   after each step; rho and Q for level l, node i}
  \item{precondition_matrices}{list of length 2 where precondition_matrices[[2]] 
                               are the pre-conditioning matrices that were used 
                               and precondition_matrices[[1]] are the combined 
                               precondition matrices}
}
}
\description{
Generalised Monte Carlo Fusion with bivariate Gaussian target
}
