% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_fusion.R
\name{parallel_fusion_uniGaussian}
\alias{parallel_fusion_uniGaussian}
\title{Preconditioned Monte Carlo Fusion [parallel]}
\usage{
parallel_fusion_uniGaussian(
  N,
  m,
  time,
  samples_to_fuse,
  means,
  sds,
  betas,
  precondition_values,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containg 
the samples for the c-th sub-posterior}

\item{means}{vector of length m, where means[c] is the mean for c-th 
sub-posterior}

\item{sds}{vector of length m, where sds[c] is the standard deviation 
for c-th sub-posterior}

\item{betas}{vector of length m, where betas[c] is the inverse temperature 
(beta) for c-th sub-posterior (can also pass in one number if 
they are all at the same inverse temperature)}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
A list with components:
\describe{
  \item{samples}{fusion samples}
  \item{rho}{acceptance rate for rho step}
  \item{Q}{acceptance rate for Q step}
  \item{rhoQ}{overall acceptance rate}
  \item{time}{run-time of fusion sampler}
  \item{rho_iterations}{number of iterations for rho step}
  \item{Q_iterations}{number of iterations for Q step}
  \item{precondition_values}{list of length 2 where precondition_values[[2]] 
                             are the pre-conditioning values that were used 
                             and precondition_values[[1]] are the combined 
                             precondition values}
}
}
\description{
Monte Carlo Fusion for sub-posteriors that are tempered Gaussian distributions
}
