% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{particle_ESS}
\alias{particle_ESS}
\title{Calculate the ESS}
\usage{
particle_ESS(log_weights)
}
\arguments{
\item{log_weights}{the logarithm of the particle weights}
}
\value{
A list with components
\describe{
 \item{log_weights}{the logarithm of the particle weights}
 \item{normalised_weights}{the normalised particle weights}
 \item{ESS}{the effective sample size of particles (the inverse of the
            sum of the weights squared)}
}
}
\description{
Calculates the ESS and the normalised weights from the logarithm of weihgts
}
\examples{
particle_ESS(log_weights = c(1000.01, 1000.02, 1000.03))
}
