% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_fusion.R
\name{Q_IS_biGaussian}
\alias{Q_IS_biGaussian}
\title{Q Importance Sampling Step}
\usage{
Q_IS_biGaussian(
  particle_set,
  m,
  time,
  mean_vec,
  sd_vec,
  corr,
  betas,
  precondition_matrices,
  inv_precondition_matrices,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{particle_set}{particles set prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}

\item{precondition_matricies}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}
}
\value{
An updated particle set
}
\description{
Q Importance Sampling weighting for bivariate Gaussian distributions
}
