% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPR_fusion.R
\name{Q_IS_BPR}
\alias{Q_IS_BPR}
\title{Q Importance Sampling Step}
\usage{
Q_IS_BPR(
  particle_set,
  m,
  time,
  dim,
  data_split,
  prior_means,
  prior_variances,
  C,
  proposal_cov,
  precondition_matrices,
  inv_precondition_matrices,
  diffusion_estimator,
  beta_NB = 10,
  gamma_NB_n_points = 2,
  seed = NULL,
  n_cores = parallel::detectCores(),
  cl = NULL,
  level = 1,
  node = 1,
  print_progress_iters = 1000
)
}
\arguments{
\item{particle_set}{particles set prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{dim}{dimension of the predictors (= p+1)}

\item{data_split}{list of length m where each item is a list of length 4 where
for c=1,...,m, data_split[[c]]$y is the vector for y responses and
data_split[[c]]$X is the design matrix for the covariates for
sub-posterior c, data_split[[c]]$full_data_count is the unique
rows of the full data set with their counts and 
data_split[[c]]$design_count is the unique rows of the design
matrix and their counts}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{C}{overall number of sub-posteriors}

\item{proposal_cov}{proposal covariance of Gaussian distribution for Fusion}

\item{precondition_matrices}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}

\item{inv_precondition_matrices}{list of length m, where inv_precondition_matrices[[c]]
is the inverse precondition matrix for sub-posterior c}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}

\item{cl}{an object of class "cluster" for parallel computation in R. If none
is passed, then one is created and used within this function}

\item{level}{indicates which level this is for the hierarchy (default 1)}

\item{node}{indicates which node this is for the hierarchy (default 1)}

\item{print_progress_iters}{number of iterations between each progress update
(default is 1000). If NULL, progress will only
be updated when importance sampling is finished}
}
\value{
An updated particle set
}
\description{
Q Importance Sampling weighting for Bayesian Poisson regression
}
