% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{Q_IS_exp_4}
\alias{Q_IS_exp_4}
\title{Q Importance Sampling Step for sub-posteriors of the form exp(-((x^4)*beta)/2)}
\usage{
Q_IS_exp_4(
  particle_set,
  m,
  time,
  mean,
  betas,
  precondition_values,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{particle_set}{particles object prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{mean}{mean value}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
An updated particle set
}
\description{
Q Importance Sampling weighting for sub-posteriors of the form exp(-((x^4)*beta)/2)
}
