% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixG_fusion.R
\name{Q_IS_mixG}
\alias{Q_IS_mixG}
\title{Q Importance Sampling Step}
\usage{
Q_IS_mixG(
  particle_set,
  m,
  time,
  n_comp,
  weights,
  means,
  sds,
  betas,
  precondition_values,
  bounds_multiplier = 1.1,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{particle_set}{particles set prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{n_comp}{integer number of components of mixture Gaussian}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gaussian}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}

\item{bounds_multiplier}{scalar value to multiply bounds by 
(should greater than or equal to 1)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
An updated particle set
}
\description{
Q Importance Sampling weighting for mixture Gaussian distributions
}
