% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_functions.R
\name{resample_indices}
\alias{resample_indices}
\title{Resampling methods}
\usage{
resample_indices(
  normalised_weights,
  method = "multi",
  n = length(normalised_weights)
)
}
\arguments{
\item{normalised_weights}{a vector of normalised weights}

\item{method}{method to be used in resampling, default is multinomial 
resampling ('multi'). Other choices are stratified ('strat'), 
systematic ('system'), residual ('resid')}

\item{n}{number of samples to resample}
}
\value{
a vector of resampled indicies
}
\description{
Resampling methods
}
