% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_functions.R
\name{resample_particle_set}
\alias{resample_particle_set}
\title{Resampling a particle set}
\usage{
resample_particle_set(
  particle_set,
  multivariate = TRUE,
  resampling_method = "multi",
  seed = NULL
)
}
\arguments{
\item{particle_set}{set of samples for x and y with corresponding weights}

\item{multivariate}{logical value indicating if the particles are multivariate
(TRUE) or not (FALSE)}

\item{resampling_method}{method to be used in resampling, default is multinomial 
resampling ('multi'). Other choices are stratified 
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{seed}{seed number - default is NULL, meaning there is no seed}
}
\value{
y parameter samples from resampled particle set
}
\description{
Resampling a particle set
}
