% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_functions.R
\name{resample_particle_x_samples}
\alias{resample_particle_x_samples}
\title{Resampling a particle set for x}
\usage{
resample_particle_x_samples(
  N = particle_set$N,
  particle_set,
  multivariate,
  step = 1,
  resampling_method = "multi",
  seed = NULL
)
}
\arguments{
\item{N}{number of samples (default is particle_set$N)}

\item{particle_set}{particle object (see initialise_particle_set)}

\item{multivariate}{logical value indicating if the particles are multivariate
(TRUE) or not (FALSE)}

\item{step}{the step of the algorithm is this is resampling for}

\item{resampling_method}{method to be used in resampling, default is multinomial 
resampling ('multi'). Other choices are stratified 
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{seed}{seed number - default is NULL, meaning there is no seed}
}
\value{
resampled particle set. Returns the same set if it has already been resampled
}
\description{
Resamples the particle set for the samples for x
}
\examples{
samples_to_fuse <- lapply(1:2, function(i) rnorm(100, 0, 1))
particles_to_fuse <- initialise_particle_sets(samples_to_fuse = samples_to_fuse,
                                              multivariate = FALSE)
precondition_values <- sapply(samples_to_fuse, var)
particles <- rho_IS_univariate(particles_to_fuse = particles_to_fuse,
                               N = 100,
                               m = 2,
                               time = 0.5,
                               precondition_values = precondition_values)
particles <- resample_particle_x_samples(particle_set = particles,
                                         multivariate = FALSE,
                                         resampling_method = 'resid',
                                         seed = 21)

}
