% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_functions.R
\name{resample_particle_y_samples}
\alias{resample_particle_y_samples}
\title{Resampling a particle set for y}
\usage{
resample_particle_y_samples(
  N = particle_set$N,
  particle_set,
  multivariate,
  resampling_method = "multi",
  seed = NULL
)
}
\arguments{
\item{N}{number of samples (default is particle_set$N)}

\item{multivariate}{logical value indicating if the particles are multivariate
(TRUE) or not (FALSE)}

\item{resampling_method}{method to be used in resampling, default is multinomial 
resampling ('multi'). Other choices are stratified 
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{seed}{seed number - default is NULL, meaning there is no seed}
}
\value{
resampled particle set. Returns the same set if it has already been resampled
}
\description{
Resamples the particle set for the samples for y
}
\examples{
particles <- create_particle(samples = rnorm(10, 0, 1), multivariate = FALSE)
particles <- resample_particle_y_samples(particle_set = particles,
                                 multivariate = FALSE,
                                 resampling_method = 'resid',
                                 seed = 21)

}
