% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_IS_univariate}
\alias{rho_IS_univariate}
\title{rho Importance Sampling Step (univariate)}
\usage{
rho_IS_univariate(particles_to_fuse, N, time, m, precondition_values)
}
\arguments{
\item{particles_to_fuse}{list of length m, where particles_to_fuse[[c]]
contains the particles for the c-th sub-posterior
(a list of particles to fuse can be initialised by
initialise_particle_sets() function)}

\item{N}{number of particles to importance sample}

\item{time}{time T for fusion algorithm}

\item{m}{number of sub-posteriors to combine}

\item{precondition_values}{precondition values for fusion}
}
\value{
A importance weighted particle set
}
\description{
Performs the importance sampling step for rho where target is univariate
}
\examples{
samples_to_fuse <- lapply(1:2, function(i) rnorm(100, 0, 1))
particles_to_fuse <- initialise_particle_sets(samples_to_fuse = samples_to_fuse,
                                              multivariate = FALSE)
precondition_values <- sapply(samples_to_fuse, var)
rho_IS_univariate(particles_to_fuse = particles_to_fuse,
                  N = 100,
                  time = 0.5,
                  m = 2,
                  precondition_values = precondition_values)
}
