% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLR_generalised_BF.R
\name{rho_j_BLR}
\alias{rho_j_BLR}
\title{rho_j Importance Sampling Step}
\usage{
rho_j_BLR(
  particle_set,
  m,
  time_mesh,
  dim,
  data_split,
  prior_means,
  prior_variances,
  C,
  precondition_matrices,
  inv_precondition_matrices,
  Lambda,
  resampling_method = "multi",
  ESS_threshold = 0.5,
  cv_location = "hypercube_centre",
  sub_posterior_means = NULL,
  adaptive_mesh = FALSE,
  adaptive_mesh_parameters = NULL,
  diffusion_estimator,
  beta_NB = 10,
  gamma_NB_n_points = 2,
  local_bounds = TRUE,
  seed = NULL,
  n_cores = parallel::detectCores(),
  cl = NULL,
  level = 1,
  node = 1,
  print_progress_iters = 1000
)
}
\arguments{
\item{particle_set}{particles set prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time_mesh}{time mesh used in Bayesian Fusion}

\item{dim}{dimension of the predictors (= p+1)}

\item{data_split}{list of length m where each item is a list of length 4 where
for c=1,...,m, data_split[[c]]$y is the vector for y responses and
data_split[[c]]$x is the design matrix for the covariates for
sub-posterior c, data_split[[c]]$full_data_count is the unique
rows of the full data set with their counts and 
data_split[[c]]$design_count is the unique rows of the design
matrix and their counts}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{C}{overall number of sub-posteriors}

\item{precondition_matrices}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}

\item{inv_precondition_matrices}{list of length m, where inv_precondition_matrices[[c]]
is the inverse precondition matrix for sub-posterior c}

\item{Lambda}{inverse of the sum of the inverse precondition matrices (which
can be computed using inverse_sum_matrices(inv_precondition_matrices))}

\item{cv_location}{string to determine what the location of the control variate
should be. Must be either 'mode' where the MLE estimator 
will be used or 'hypercube_centre' (default) to use the centre
of the simulated hypercube}

\item{sub_posterior_means}{matrix with m rows and d columns, where sub_posterior_means[c,]
is the sub-posterior mean of sub-posterior c}

\item{adaptive_mesh}{logical value to indicate if an adaptive mesh is used
(default is FALSE)}

\item{adaptive_mesh_parameters}{list of parameters used for adaptive mesh}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{local_bounds}{logical value indicating if local bounds for the phi function
are used (default is TRUE)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}

\item{cl}{an object of class "cluster" for parallel computation in R. If none
is passed, then one is created and used within this function}

\item{level}{indicates which level this is for the hierarchy (default 1)}

\item{node}{indicates which node this is for the hierarchy (default 1)}

\item{print_progress_iters}{number of iterations between each progress update
(default is 1000). If NULL, progress will only
be updated when importance sampling is finished}
}
\value{
A list with components:
\describe{
  \item{particle_set}{updated particle set after the iterative rho_j steps}
  \item{proposed_samples}{proposal samples for the last time step}
  \item{time}{elapsed time of each step of the algorithm}
  \item{ESS}{effective sample size of the particles after each step}
  \item{CESS}{conditional effective sample size of the particles after each step}
  \item{resampled}{boolean value to indicate if particles were resampled
                   after each time step}
  \item{E_nu_j}{Approximation of the average variation of the trajectories
                at each time step}
}
}
\description{
rho_j Importance Sampling weighting for Bayesian logistic regression
}
