% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixG_sampler.R
\name{rmix_norm}
\alias{rmix_norm}
\title{Simulate exactly a mixture Gaussian distribution}
\usage{
rmix_norm(N, weights, means, sds)
}
\arguments{
\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gassuan}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{n}{number of samples required}
}
\value{
samples from a mixture Gaussian distribution
}
\description{
Simulate exactly a mixture Gaussian distribution using component sampling
}
\examples{
rmix_norm(N = 10000, weights = c(0.4, 0.6), means = c(-3, 9), sds = c(2, 0.4))

}
