% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixG_sampler.R
\name{rmix_norm_tempered_rejection_sampler}
\alias{rmix_norm_tempered_rejection_sampler}
\title{Rejection sampler for tempered mixture Gaussian}
\usage{
rmix_norm_tempered_rejection_sampler(
  N,
  weights,
  means,
  sds,
  beta,
  proposal_sds,
  dominating_M
)
}
\arguments{
\item{N}{number of samples}

\item{weights}{vector: weights of mixture Gaussian}

\item{means}{vector: means of mixture Gassuan}

\item{sds}{vector: st.devs of mixture Gaussian}

\item{beta}{real value between 0 and 1}

\item{proposal_sds}{vector: st.devs of proposal mixture Gaussian}

\item{dominating_M}{constant M to bound the target density}
}
\value{
samples from tempered target density
}
\description{
Sample from tempered target using rejection sampling with mixture Gaussian proposals
}
