% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_sampler.R
\name{rnorm_tempered}
\alias{rnorm_tempered}
\title{Tempered Gaussian Distribution Sampler}
\usage{
rnorm_tempered(N, mean, sd, beta)
}
\arguments{
\item{mean}{mean value}

\item{sd}{standard deviation value}

\item{beta}{real value}

\item{x}{real value}
}
\value{
N samples from a tempered Gaussian distribution
}
\description{
Samples from a tempered Gaussian distribution
}
\examples{
# standard normal
curve(dnorm(x), -5, 5)
# standard normal to the power 1/2
curve(dnorm_tempered(x, beta = 1/2), -5, 5, add = T, col = 'blue')
samples <- rnorm_tempered(N = 10000, mean = 0, sd = 1, beta = 1/2)
lines(density(samples), col = 'red')

}
