% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_wise_subtraction}
\alias{row_wise_subtraction}
\title{Row-wise subtraction of a vector to rows of a matrix}
\usage{
row_wise_subtraction(X, vect)
}
\arguments{
\item{X}{matrix (n x m)}

\item{vect}{vector of length m}
}
\value{
an (n x m) matrix, Y, where Y[i,] = X[i,]-vect
}
\description{
Calculates the subtraction of a vector to each row of a matrix
}
\examples{
X <- matrix(c(1,2,3,4,5,6,7,8), nrow = 4, ncol = 2, byrow = T)
row_wise_subtraction(X = X, vect = c(1,2))
}
