% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_sampler.R
\name{sample_exp_4}
\alias{sample_exp_4}
\title{Rejection sampler for tempered target exp(-(beta*(x-mean)^4)/2)}
\usage{
sample_exp_4(N, mean = 0, proposal_mean, proposal_sd, dominating_M, beta)
}
\arguments{
\item{N}{number of samples}

\item{mean}{mean value, defaults to 0}

\item{proposal_mean}{mean of proposal density for rejection sampler}

\item{proposal_sd}{st.dev of proposal density for rejection sampler}

\item{beta}{real value between 0 and 1}

\item{dominiating_M}{constant M to bound the target density}
}
\value{
samples from tempered target density
}
\description{
Sample from tempered target using rejection sampling with normal proposals
}
\examples{
dominating_dnorm <- function(x) 1.3*dnorm(x, mean = 0, sd = 1)
curve(dominating_dnorm(x), -5, 5, col = 'red')
curve(dnorm(x, mean = 0, sd = 1), -5, 5, add = T, col = 'green')
curve(exp_4_density(x, 1/2), -5, 5, add = T, col = 'blue')
sample_from_fc(N = 100000, proposal_mean = 0, proposal_sd = 1, dominating_M = 1.3)

}
