% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaled_distance}
\alias{scaled_distance}
\title{Scaled distance between two vectors}
\usage{
scaled_distance(x, y, matrix)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{matrix}{matrix}
}
\value{
the scaled distance between vectors x and y with matrix
}
\description{
Calculates the scaled distance between two vectors, i.e. calculates the norm of matrix*(x-y)
If matrix == identity matrix, this is just the Euclidean distance
}
\examples{
x <- c(0.3, 0.2, 0.5, 1.2)
y <- c(-0.5, 0.8, 1.4, 0.9)
scaled_distance(x, y, diag(1, 4))
# should equal to the Euclidean distance:
sqrt(0.8^2 + 0.6^2 + 0.9^2 + 0.3^2)
}
