% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectral_radius}
\alias{spectral_radius}
\title{Spectral radius of a symmetric matrix}
\usage{
spectral_radius(A)
}
\arguments{
\item{A}{matrix}
}
\value{
The spectral radius (largest absolute eigenvalue) of A
}
\description{
Calculates the spectral radius of a symmetric matrix A (the largest absolute eigenvalue)
}
\examples{
# symmetric matrix
# should equal 2.5
spectral_radius(matrix(c(2, 0.5, 0.5, 2), nrow = 2, ncol = 2))
# non-symmetric matrix
# should equal 10
spectral_radius(matrix(c(9, -1, 2, -2, 8, 4, 1, 1, 8), nrow = 3, ncol = 3, byrow = T))
}
