% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_mean_univariate}
\alias{weighted_mean_univariate}
\title{Calculate the weighted mean (univariate)}
\usage{
weighted_mean_univariate(x, weights)
}
\arguments{
\item{x}{vector of values}

\item{weights}{vector of weights}
}
\value{
the weighted mean of x
}
\description{
Calculation of weighted mean when the target is univariate
}
\examples{
x <- c(0.4, 0.2, 0.5, 0.9, 1.4)
w <- c(3, 2, 5, 1, 2)
weighted_mean_univariate(x = x,
                         weights = w)
# returns the same using weighted.mean(x, w) function in R
weighted.mean(x, w)

weighted_mean_univariate(x = x,
                         weights = rep(1, 5))
#returns the same using standard mean(x) function in R
mean(x)
}
