% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_trajectory_variation_multivariate}
\alias{weighted_trajectory_variation_multivariate}
\title{Calculate approximation to expectation of nu_j (multivariate)}
\usage{
weighted_trajectory_variation_multivariate(
  x_samples,
  sub_posterior_means,
  inv_precondition_matrices
)
}
\arguments{
\item{sub_posterior_means}{list of length C, where sub_posterior_mean[[c]] is the
sub_posterior_mean of the cth sub-posterior}

\item{inv_precondition_matrices}{list of length m of inverse 
preconditioning matrices}

\item{list}{where x_samples[[i]] ith collection of the C trajectories}
}
\value{
the approximated expectation of nu_j
}
\description{
Calculation of the scaled/weighted average variation of the C trajectories
with respect to their individual sub-posterior means
}
\examples{
N <- 10
C <- 4
d <- 3
x_samples <- lapply(1:N, function(i) mvrnormArma(C, rep(0,d), diag(1,d)))
sub_posterior_means <- mvrnormArma(C, rep(0,d), diag(1,d))
precond <- lapply(1:C, function(c) diag(c, d))
inv_precond <- lapply(precond, solve)
weighted_trajectory_variation_multivariate(x_samples = x_samples,
                                           sub_posterior_means = sub_posterior_means,
                                           inv_precondition_matrices = inv_precond)
# should be equal to the result of this:
sum(sapply(1:N, function(i) {
  sum(sapply(1:C, function(c) {
    diff <- x_samples[[i]][c,]-sub_posterior_means[c,]
    return(t(diff) \%*\% inv_precond[[c]] \%*\% diff)
  }))/C
}))/N
}
