% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_trajectory_variation_univariate}
\alias{weighted_trajectory_variation_univariate}
\title{Calculate approximation to expectation of nu_j (univariate)}
\usage{
weighted_trajectory_variation_univariate(
  x_samples,
  sub_posterior_means,
  precondition_values
)
}
\arguments{
\item{sub_posterior_means}{list of length C, where sub_posterior_mean[[c]] is the
sub_posterior_mean of the cth sub-posterior}

\item{precondition_values}{precondition values associated to each sub-posterior}

\item{list}{where x_samples[[i]] ith collection of the C trajectories}
}
\value{
the approximated expectation of nu_j
}
\description{
Calculation of the scaled/weighted average variation of the C trajectories
with respect to their individual sub-posterior means
}
\examples{
# x_samples has 5 samples and C=4
x <- rnorm(4, 0, 1)
x_samples <- rep(list(x), 5)
precondition_vals <- c(1, 2, 3, 4)
# compute x_means (and take these as the sub-posterior means)
x_mean <- weighted_mean_univariate(x = x,
                                   weights = 1/precondition_vals)
sub_posterior_means <- rep(list(x_mean), 4)
# this should be 3 times the example for weighted_variance_univariate()
weighted_trajectory_variation_univariate(x_samples = x_samples,
                                         sub_posterior_means = sub_posterior_means,
                                         precondition_values = precondition_vals)
3*weighted_variance_univariate(x = x,
                               x_mean = x_mean,
                               precondition_values = precondition_vals)
}
