% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_variance_univariate}
\alias{weighted_variance_univariate}
\title{Calculate the variance of numbers (univariate)}
\usage{
weighted_variance_univariate(x, x_mean, precondition_values)
}
\arguments{
\item{x}{vector of sampled sub-posterior values}

\item{x_mean}{weighted mean of sampled sub-posterior values}

\item{precondition_values}{precondition values associated to each sub-posterior}
}
\value{
the weighted variance of numbers
}
\description{
Calculation of the weighted variance of numbers
}
\examples{
x <- rnorm(4, 0, 1)
precondition_vals <- c(1, 2, 3, 4)
x_mean <- weighted_mean_univariate(x = x,
                                   weights = 1/precondition_vals)
weighted_variance_univariate(x = x,
                             x_mean = x_mean,
                             precondition_values = precondition_vals)
}
