% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_functions.R
\name{create_particle}
\alias{create_particle}
\title{Create a particle object}
\usage{
create_particle(samples, multivariate)
}
\arguments{
\item{multivariate}{logical value indicating if the samples are multivariate
(TRUE) or not (FALSE)}

\item{samples_to_fuse}{a list of samples (vector if univariate, matrix is multivariate)
that you wish to perform fusion with}
}
\value{
A particle environment with components
\describe{
  \item{y_samples}{samples for y in particle set (initialised as the samples given)}
  \item{x_samples}{a list where x_samples[[i]] is the ith x sample for in 
                   particle set (all initialised as NA)}
  \item{x_mean}{the corresponding means for x_samples (initialised as NA)}
  \item{log_weights}{associated logarithm of the weights (initiliased as
                     the logarithm of 1/number of samples)}
  \item{normalised_weights}{associated normalised weights (initialised as 
                            1/number of samples)}
  \item{ESS}{effective sample size of particles (initialised as the 
             number of samples)}
  \item{CESS}{conditional effective sample size of particles after rho and Q step
              (initialised as NA for both)}
  \item{resampled}{logical value to idicate if particles have been resampled
                   after rho and Q step (initialised as FALSE for rho and TRUE for Q)}
  \item{N}{Number of particles}
}
}
\description{
Creates an R environment to contain the particles for SMC Fusion
}
\examples{
p <- create_particle(samples = rnorm(10, 0, 1), multivariate = FALSE)

}
