% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{ea_exp_4_DL}
\alias{ea_exp_4_DL}
\title{Exact Algorithm for Langevin diffusion with pi = exp(-(beta*(x-mean)^4)/2)}
\usage{
ea_exp_4_DL(
  N,
  input_samples,
  time,
  mean,
  beta,
  precondition,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2
)
}
\arguments{
\item{N}{number of samples}

\item{input_samples}{input samples for the algorithm distributed according to
pi = exp(-(beta*(x-mean)^4)/2)}

\item{time}{time T for Exact Algorithm}

\item{mean}{mean value}

\item{beta}{real value}

\item{precondition}{precondition value (i.e the covariance for 
the Langevin diffusion)}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}
}
\value{
end points of the Exact Algorithm which should also be distributed
        according to pi = exp(-(beta*(x-mean)^4)/2)
}
\description{
Exact Algorithm with pi = exp(-(beta*(x-mean)^4)/2)
}
