% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{ea_exp_4_DL_PT}
\alias{ea_exp_4_DL_PT}
\title{Diffusion probability for the Exact Algorithm for langevin diffusion 
with pi = exp(-(beta*(x-mean)^4)/2)}
\usage{
ea_exp_4_DL_PT(
  x0,
  y,
  s,
  t,
  mean,
  beta,
  precondition,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  logarithm
)
}
\arguments{
\item{x0}{start value}

\item{y}{end value}

\item{s}{start time}

\item{t}{end time}

\item{mean}{mean value}

\item{beta}{real value}

\item{precondition}{precondition value (i.e the covariance for 
the Langevin diffusion)}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poission estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{logarithm}{logical value to determine if log probability is 
returned (TRUE) or not (FALSE)}

\item{K}{lower bound of the phi function}
}
\value{
acceptance probability of simulating langevin diffusion with pi = 
        exp(-(beta*(x-mean)^4)/2)
}
\description{
Evaluate probability of langevin diffusion using the Exact Algorithm with pi = 
exp(-(beta*(x-mean)^4)/2) using Poisson thinning
}
\examples{
mu <- 0.435
beta <- 0.583
precondition <- 1.243
# Poisson estimator
ea_exp_4_DL_PT(x0 = 0,
               y = 0.23,
               s = 0,
               t = 1,
               mean = mu,
               beta = beta,
               precondition = precondition,
               logarithm = TRUE)
# NB estimator
ea_exp_4_DL_PT(x0 = 0,
               y = 0.23,
               s = 0,
               t = 1,
               mean = mu,
               beta = beta,
               precondition = precondition,
               diffusion_estimator = 'NB',
               logarithm = TRUE)

}
