% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_4_fusion.R
\name{fusion_exp_4}
\alias{fusion_exp_4}
\title{Generalised Monte Carlo Fusion (rejection sampling) [on a single core]}
\usage{
fusion_exp_4(N, m, time, samples_to_fuse, mean, betas, precondition_values)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containing 
the samples for the c-th sub-posterior}

\item{mean}{mean value of target distribution}

\item{betas}{vector of length m, where betas[c] is the inverse temperature 
(beta) for c-th sub-posterior (can also pass in one number if 
they are all at the same inverse temperature)}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}
}
\value{
A list with components:
\describe{
  \item{samples}{fusion samples}
  \item{iters_rho}{number of iterations for rho step}
  \item{iters_Q}{number of iterations for Q step}
}
}
\description{
Generalised Monte Carlo Fusion with target of the form exp(-((x^4)*beta)/2)
}
