% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_fusion.R
\name{fusion_uniGaussian}
\alias{fusion_uniGaussian}
\title{Preconditioned Monte Carlo Fusion [on a single core]}
\usage{
fusion_uniGaussian(
  N,
  m,
  time,
  samples_to_fuse,
  means,
  sds,
  betas,
  precondition_values
)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[c] containg 
the samples for the c-th sub-posterior}

\item{means}{vector of length m, where means[c] is the mean for c-th 
sub-posterior}

\item{sds}{vector of length m, where sds[c] is the standard deviation 
for c-th sub-posterior}

\item{betas}{vector of length m, where betas[c] is the inverse temperature 
(beta) for c-th sub-posterior (can also pass in one number if 
they are all at the same inverse temperature)}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}
}
\value{
A list with components:
\describe{
  \item{samples}{fusion samples}
  \item{iters_rho}{number of iterations for rho step}
  \item{iters_Q}{number of iterations for Q step}
}
}
\description{
Monte Carlo Fusion for sub-posteriors that are tempered Gaussian distributions
}
