% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_functions.R
\name{initialise_particle_sets}
\alias{initialise_particle_sets}
\title{Initialise particle sets from a list of samples}
\usage{
initialise_particle_sets(samples_to_fuse, multivariate)
}
\arguments{
\item{samples_to_fuse}{a list of samples that you wish to perform fusion with}

\item{multivariate}{logical value indicating if the samples are multivariate
(TRUE) or not (FALSE)}
}
\value{
A list with components
\describe{
  \item{y_samples}{samples for y in particle set (initialised as the samples given)}
  \item{x_samples}{a list where x_samples[[i]] is the ith x sample for in 
                   particle set (initialised as NA)}
  \item{x_mean}{the corresponding means for x_samples (initialised as NA)}
  \item{log_weights}{associated logarithm of the weights (initiliased as
                     the logarithm of 1 / number of samples)}
  \item{normalised_weights}{associated normalised weights (initialised as 
                            1 / number of samples)}
  \item{ESS}{effective sample size of particles (initialised as the 
             number of samples)}
  \item{CESS}{conditional effective sample size of particles (initialised 
              as the number of samples)}
  \item{resampled}{logical value to idicate if particles have been resampled
                   (initialised as TRUE)}
  \item{N}{Number of particles}
}
}
\description{
Function to intialise particle sets from a list of samples
}
\examples{
samples <- rnorm(100, 0, 1)
initialise_particle_sets(samples_to_fuse = samples, multivariate = FALSE)

}
