% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLR_fusion.R
\name{obtain_LR_MLE}
\alias{obtain_LR_MLE}
\title{#' @export
surrounding_hypercube_vertices <- function(hypercube_vertices) {
  bounds <- lapply(1:ncol(hypercube_vertices), function(d) c(min(hypercube_vertices[,d]), max(hypercube_vertices[,d])))
  vertices <- as.matrix(expand.grid(bounds))
  colnames(vertices) <- c()
  return(vertices)
}}
\usage{
obtain_LR_MLE(dim, data)
}
\description{
#' @export
surrounding_hypercube_vertices <- function(hypercube_vertices) {
  bounds <- lapply(1:ncol(hypercube_vertices), function(d) c(min(hypercube_vertices[,d]), max(hypercube_vertices[,d])))
  vertices <- as.matrix(expand.grid(bounds))
  colnames(vertices) <- c()
  return(vertices)
}
}
