% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_fusion.R
\name{parallel_fusion_biGaussian}
\alias{parallel_fusion_biGaussian}
\title{Generalised Monte Carlo Fusion (rejection sampling) [parallel]}
\usage{
parallel_fusion_biGaussian(
  N,
  m,
  time,
  samples_to_fuse,
  mean_vec,
  sd_vec,
  corr,
  betas,
  precondition_matrices,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{samples_to_fuse}{list of length m, where samples_to_fuse[[c]]
contains the samples for the c-th sub-posterior}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}

\item{precondition_matricies}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}
}
\value{
A list with components:
\describe{
 \item{samples}{fusion samples}
 \item{rho}{acceptance rate for rho step}
 \item{Q}{acceptance rate for Q step}
 \item{rhoQ}{overall acceptance rate}
 \item{time}{run-time of fusion sampler}
 \item{rho_iterations}{number of iterations for rho step}
 \item{Q_iterations}{number of iterations for Q step}
 \item{precondition_matrices}{list of length 2 where precondition_matrices[[2]] 
                              are the pre-conditioning matrices that were used 
                              and precondition_matrices[[1]] are the combined 
                              precondition matrices}
}
}
\description{
Generalised Monte Carlo Fusion with bivariate Gaussian target
}
