% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLR_fusion.R
\name{parallel_fusion_SMC_BLR}
\alias{parallel_fusion_SMC_BLR}
\title{Generalised Monte Carlo Fusion [parallel]}
\usage{
parallel_fusion_SMC_BLR(
  particles_to_fuse,
  N,
  m,
  time,
  dim,
  data_split,
  prior_means,
  prior_variances,
  C,
  precondition_matrices,
  resampling_method = "multi",
  ESS_threshold = 0.5,
  cv_location = "hypercube_centre",
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  local_bounds = TRUE,
  seed = NULL,
  n_cores = parallel::detectCores(),
  cl = NULL,
  level = 1,
  node = 1,
  print_progress_iters = 1000
)
}
\arguments{
\item{particles_to_fuse}{list of length m, where particles_to_fuse[c] contains
the particles for the c-th sub-posterior. Can
initialise a this from list of sub-posterior samples
by using the initialise_particle_sets function}

\item{N}{number of samples}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{dim}{dimension of the predictors (= p+1)}

\item{data_split}{list of length m where each item is a list of length 4 where
for c=1,...,m, data_split[[c]]$y is the vector for y responses and
data_split[[c]]$X is the design matrix for the covariates for
sub-posterior c, data_split[[c]]$full_data_count is the unique
rows of the full data set with their counts and 
data_split[[c]]$design_count is the unique rows of the design
matrix and their counts}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{C}{overall number of sub-posteriors}

\item{precondition_matrices}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}

\item{resampling_method}{method to be used in resampling, default is multinomial
resampling ('multi'). Other choices are stratified
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{ESS_threshold}{number between 0 and 1 defining the proportion of the
number of samples that ESS needs to be lower than for
resampling (i.e. resampling is carried out only when
ESS < N*ESS_threshold)}

\item{cv_location}{string to determine what the location of the control variate
should be. Must be either 'mode' where the MLE estimator 
will be used or 'hypercube_centre' (default) to use the centre
of the simulated hypercube}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{local_bounds}{logical value indicating if local bounds for the phi function
are used (default is TRUE)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}

\item{cl}{an object of class "cluster" for parallel computation in R. If none
is passed, then one is created and used within this function}

\item{level}{indicates which level this is for the hierarchy (default 1)}

\item{node}{indicates which node this is for the hierarchy (default 1)}

\item{print_progress_iters}{number of iterations between each progress update
(default is 1000). If NULL, progress will only
be updated when importance sampling is finished}
}
\value{
A list with components:
\describe{
  \item{particles}{particles returned from fusion sampler}
  \item{proposed_samples}{proposal samples from fusion sampler}
  \item{time}{run-time of fusion sampler}
  \item{ESS}{effective sample size of the particles after each step}
  \item{CESS}{conditional effective sample size of the particles after each step}
  \item{resampled}{boolean value to indicate if particles were resampled
                   after each time step}
  \item{precondition_matrices}{list of length 2 where precondition_matrices[[2]]
                               are the pre-conditioning matrices that were used
                               and precondition_matrices[[1]] are the combined
                               precondition matrices}
  \item{combined_data}{combined data for the fusion density}
}
}
\description{
Generalised Monte Carlo Fusion for Bayesian Logistic Regression
}
