% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_Gaussian_fusion.R
\name{Q_IS_uniGaussian}
\alias{Q_IS_uniGaussian}
\title{Q Importance Sampling Step for tempered bivariate Gaussian distributions}
\usage{
Q_IS_uniGaussian(
  particle_set,
  m,
  time,
  means,
  sds,
  betas,
  precondition_values,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{particle_set}{particles set prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time}{time T for fusion algorithm}

\item{means}{vector of length m, where means[c] is the mean for c-th 
sub-posterior}

\item{sds}{vector of length m, where sds[c] is the standard deviation 
for c-th sub-posterior}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{precondition_values}{vector of length m, where precondition_values[c]
is the precondition value for sub-posterior c}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
An updated particle set
}
\description{
Q Importance Sampling weighting for tempered bivariate Gaussian distributions
}
