% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_Gaussian_generalised_BF.R
\name{rho_j_biGaussian}
\alias{rho_j_biGaussian}
\title{rho_j Importance Sampling Step}
\usage{
rho_j_biGaussian(
  particle_set,
  m,
  time_mesh,
  mean_vec,
  sd_vec,
  corr,
  betas,
  precondition_matrices,
  inv_precondition_matrices,
  Lambda,
  resampling_method = "multi",
  ESS_threshold = 0.5,
  diffusion_estimator = "Poisson",
  beta_NB = 10,
  gamma_NB_n_points = 2,
  seed = NULL,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{particle_set}{particles set prior to Q importance sampling step}

\item{m}{number of sub-posteriors to combine}

\item{time_mesh}{time mesh used in Bayesian Fusion}

\item{mean_vec}{vector of length 2 for mean}

\item{sd_vec}{vector of length 2 for standard deviation}

\item{corr}{correlation value between component 1 and component 2}

\item{betas}{vector of length c, where betas[c] is the inverse temperature 
value for c-th posterior}

\item{precondition_matrices}{list of length m, where precondition_matrices[[c]]
is the precondition matrix for sub-posterior c}

\item{inv_precondition_matrices}{list of length m, where inv_precondition_matrices[[c]]
is the inverse precondition matrix for sub-posterior c}

\item{Lambda}{inverse of the sum of the inverse precondition matrices (which
can be computed using inverse_sum_matrices(inv_precondition_matrices))}

\item{resampling_method}{method to be used in resampling, default is multinomial
resampling ('multi'). Other choices are stratified
resampling ('strat'), systematic resampling ('system'),
residual resampling ('resid')}

\item{ESS_threshold}{number between 0 and 1 defining the proportion of the
number of samples that ESS needs to be lower than for
resampling (i.e. resampling is carried out only when
ESS < N*ESS_threshold)}

\item{diffusion_estimator}{choice of unbiased estimator for the Exact Algorithm
between "Poisson" (default) for Poisson estimator
and "NB" for Negative Binomial estimator}

\item{beta_NB}{beta parameter for Negative Binomial estimator (default 10)}

\item{gamma_NB_n_points}{number of points used in the trapezoidal estimation
of the integral found in the mean of the negative
binomial estimator (default is 2)}

\item{seed}{seed number - default is NULL, meaning there is no seed}

\item{n_cores}{number of cores to use}
}
\value{
A list with components:
\describe{
  \item{particle_set}{updated particle set after the iterative rho_j steps}
  \item{proposed_samples}{proposal samples for the last time step}
  \item{ESS}{effective sample size of the particles after each step}
  \item{CESS}{conditional effective sample size of the particles after each step}
  \item{resampled}{boolean value to indicate if particles were resampled
                   after each time step}
}
}
\description{
rho_j Importance Sampling weighting for bivariate Gaussian distributions
}
