% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_trajectory_variation_univariate}
\alias{weighted_trajectory_variation_univariate}
\title{Calculate approximation to expectation of nu_j (univariate)}
\usage{
weighted_trajectory_variation_univariate(
  x_samples,
  sub_posterior_means,
  precondition_values
)
}
\arguments{
\item{sub_posterior_means}{vector of length C of sub-posterior means}

\item{precondition_values}{precondition values associated to each sub-posterior}

\item{list}{where x_samples[[i]] ith collection of the C trajectories}
}
\value{
the approximated expectation of nu_j
}
\description{
Calculation of the scaled/weighted average variation of the C trajectories
with respect to their individual sub-posterior means
}
\examples{
# x_samples has 5 samples and C=4
N <- 10
C <- 4
x_samples <- lapply(1:N, function(i) rnorm(C))
sub_posterior_means <- rnorm(C)
precond <- 1:C
weighted_trajectory_variation_univariate(x_samples = x_samples,
                                         sub_posterior_means = sub_posterior_means,
                                         precondition_values = precond)
# should be equal to the result of this:
sum(sapply(1:N, function(i) {
  sum((((x_samples[[i]]-sub_posterior_means)^2)/precond))/C
}))/N
}
